+++ shaders/lib/lighting.glsl
@@ -82,10 +82,6 @@ varying float skyBrightnessMult;
 	
 #endif
 
-#ifdef HANDHELD_LIGHT_ENABLED
-	varying float handLightBrightness;
-#endif
-
 
 
 
@@ -138,7 +134,7 @@ vec3 getLightingBrightnesses(vec2 lmcoord) {
 			
 			vec3 offsetShadowPos = shadowPos;
 			vec2 noise = randomVec2(rngStart);
-			offsetShadowPos.xy += noise * offsetMult * 0.25;
+			offsetShadowPos.xy += noise * offsetMult * 0.15;
 			
 			#if SHADOW_FILTERING == 0
 				
@@ -151,6 +147,8 @@ vec3 getLightingBrightnesses(vec2 lmcoord) {
 				
 				// filtered
 				// tactic: just absorb the shadow acne and average it out, then multiply and clamp to get back to 1.0
+				// actually I don't think that's how this works
+				// the problem is that the offset pos goes inside the block half the time, which gets counted as in shadow when the lightDot is low
 				for (int i = 0; i < SHADOW_OFFSET_COUNT; i++) {
 					if (texture2D(shadowtex0, offsetShadowPos.xy + SHADOW_OFFSETS[i].xy * offsetMult).r >= offsetShadowPos.z) {
 						float currentShadowWeight = SHADOW_OFFSETS[i].z;
@@ -158,7 +156,7 @@ vec3 getLightingBrightnesses(vec2 lmcoord) {
 					}
 				}
 				skyBrightness /= SHADOW_OFFSET_WEIGHTS_TOTAL;
-				const float shadowMult1 = 1.0; // for when lightDot is 1.0 (sun is directly facing surface)
+				const float shadowMult1 = 1.4; // for when lightDot is 1.0 (sun is directly facing surface)
 				const float shadowMult2 = 2.2; // for when lightDot is 0.0 (sun is angled relative to surface)
 				skyBrightness = min(skyBrightness * (shadowMult2 - skyBrightnessMult * (shadowMult2 - shadowMult1)), 1.0);
 				
@@ -175,13 +173,7 @@ vec3 getLightingBrightnesses(vec2 lmcoord) {
 	
 	
 	float blockBrightness = lmcoord.x;
-	#ifdef HANDHELD_LIGHT_ENABLED
-		blockBrightness = max(blockBrightness, handLightBrightness);
-	#endif
-	
 	float ambientBrightness = lmcoord.y;
-	
-	
 	return vec3(blockBrightness, skyBrightness, ambientBrightness);
 }
 
@@ -212,7 +204,6 @@ void doPreLighting() {
 		
 		if (lightDot > 0.0) { // vertex is facing towards the sky
 			vec4 viewPos = gl_ModelViewMatrix * gl_Vertex;
-			vec4 worldPos = gbufferModelViewInverse * viewPos;
 			#if SHADOW_FILTERING == 0
 				shadowPos = getShadowPos(viewPos, lightDot);
 			#else
@@ -229,12 +220,11 @@ void doPreLighting() {
 	
 	
 	#ifdef HANDHELD_LIGHT_ENABLED
-		handLightBrightness = 0.0;
 		float depth = estimateDepthVSH();
 		if (depth <= HANDHELD_LIGHT_DISTANCE) {
-			handLightBrightness = max(1.0 - depth / HANDHELD_LIGHT_DISTANCE, 0.0);
-			handLightBrightness = handLightBrightness;
+			float handLightBrightness = max(1.0 - depth / HANDHELD_LIGHT_DISTANCE, 0.0);
 			handLightBrightness *= heldBlockLightValue / 15.0 * HANDHELD_LIGHT_BRIGHTNESS;
+			lmcoord.x = max(lmcoord.x, handLightBrightness);
 		}
 	#endif