+++ shaders/main_files/gbuffers_clouds.glsl
@@ -1,7 +1,23 @@
+// defines
+
+#if defined BLOOM_ENABLED && defined NORMALS_NEEDED
+	#define BLOOM_AND_NORMALS
+#endif
+
+// transfers
+
 varying vec2 texcoord;
 flat float glcolor;
 
-#include "/lib/fog.glsl"
+#ifdef NORMALS_NEEDED
+	varying vec3 normal;
+#endif
+
+// includes
+
+#ifdef FOG_ENABLED
+	#include "/lib/fog.glsl"
+#endif
 
 
 
@@ -36,9 +52,16 @@ void main() {
 		color.rgb = debugOutput;
 	#endif
 	gl_FragData[0] = color;
-	#ifdef BLOOM_ENABLED
+	#ifdef BLOOM_AND_NORMALS
+		/* DRAWBUFFERS:024 */
+		gl_FragData[1] = colorForBloom;
+		gl_FragData[2] = vec4(normal, 1.0);
+	#elif defined BLOOM_ENABLED
 		/* DRAWBUFFERS:02 */
 		gl_FragData[1] = colorForBloom;
+	#elif defined NORMALS_NEEDED
+		/* DRAWBUFFERS:04 */
+		gl_FragData[1] = vec4(normal, 1.0);
 	#endif
 }
 
@@ -63,6 +86,10 @@ void main() {
 	#endif
 	
 	glcolor = gl_Color.r;
+	
+	#ifdef NORMALS_NEEDED
+		normal = gl_NormalMatrix * gl_Normal;
+	#endif
 	
 }