+++ shaders/main_files/gbuffers_hand.glsl
@@ -1,10 +1,24 @@
+// defines
+
+#undef SHADOWS_ENABLED
+
+#if defined BLOOM_ENABLED && defined NORMALS_NEEDED
+	#define BLOOM_AND_NORMALS
+#endif
+
+// transfers
+
 varying vec2 texcoord;
 varying vec2 lmcoord;
 varying vec4 glcolor;
 
-#undef SHADOWS_ENABLED
+#ifdef NORMALS_NEEDED
+	varying vec3 normal;
+#endif
 
-#include "../lib/lighting.glsl"
+// includes
+
+#include "/lib/lighting.glsl"
 
 
 
@@ -41,9 +55,16 @@ void main() {
 		color = vec4(debugOutput, 1.0);
 	#endif
 	gl_FragData[0] = color;
-	#ifdef BLOOM_ENABLED
+	#ifdef BLOOM_AND_NORMALS
+		/* DRAWBUFFERS:024 */
+		gl_FragData[1] = colorForBloom;
+		gl_FragData[2] = vec4(normal, 1.0);
+	#elif defined BLOOM_ENABLED
 		/* DRAWBUFFERS:02 */
 		gl_FragData[1] = colorForBloom;
+	#elif defined NORMALS_NEEDED
+		/* DRAWBUFFERS:04 */
+		gl_FragData[1] = vec4(normal, 1.0);
 	#endif
 }
 
@@ -67,6 +88,10 @@ void main() {
 	
 	
 	glcolor = gl_Color;
+	
+	#ifdef NORMALS_NEEDED
+		normal = gl_NormalMatrix * gl_Normal;
+	#endif
 	
 	
 	doPreLighting();